Update categories and eligibilities for DCYF.
This adds a migration that makes a large number of changes to categories
and eligibilities, to match the changes requested by DCYF. Each group of
categories and elgibilities is handled a little bit differently, but
they roughly fall into one of the following patterns:

- Create a new category/eligibility
- Rename a category/eligibility
- Mass migrate a number of resources and services from a set of old
  categories/eligibilities to a new one
- Sometimes delete the old categories/eligibilities when doing said mass
  migration, though in some cases preserve the old ones

This was implemented as a pure data migration that does not change the
schema. Since data migrations must be agnostic to the actual models, we
cannot use any of the model code, and instead we construct a number of
raw SQL queries. These queries have been factored into independent
methods so that the main migration code can be read top to bottom with
high-level method calls describing the exact changes being made.