Add `boosted_category_id` column and add to Algolia index. (#748)
This adds a `boosted_category_id` column to Services, which is a
nullable foreign key to the Categories table. This is intended to be
used by the frontend code through Algolia as a possible target of the
optional filters feature of Algolia, allowing us to boost search
rankings for specific services on specific pages.

This column is added to the Algolia index as `boosted_category`, which
holds the string name of the category, and we also add it to the
`attributesForFaceting` list so that it can be used for optional
filters.

Finally, we change the Algolia index ranking order from its default by
placing `geo` after `filters`. This is needed in order for optional
filters to have higher precedence than geolocation.