Update URL for PostgreSQL apt packages. (#763)
Our CI builds are failing because the PostgreSQL apt packages can't be
downloaded anymore. This is probably due to the fact that our Docker
image ultimately uses Ubuntu 20.04 Focal as the base OS, which is now
out of long-term support, and so PostgreSQL doesn't support it anymore.

It looks like the packages are still available, but under a different
URL with "archive" in the name. This updates the Dockerfile to point to
that new URL.